@@ -49,7 +49,7 @@
     "format": "./node_modules/.bin/nx format:write",
     "format:write": "./node_modules/.bin/nx format:write",
     "format:check": "./node_modules/.bin/nx format:check",
-    "update": "ng update @nrwl/schematics",
+    "update": "ng update @nrwl/workspace",
     "update:check": "ng update",
     "workspace-schematic": "./node_modules/.bin/nx workspace-schematic",
     "dep-graph": "./node_modules/.bin/nx dep-graph",
@@ -59,38 +59,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.0",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "^7.2.0",
-    "@angular/material": "^7.2.1",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
-    "@ngrx/effects": "7.0.0",
-    "@ngrx/router-store": "7.0.0",
-    "@ngrx/store": "7.0.0",
-    "@nrwl/nx": "7.3.0",
+    "@angular/animations": "^8.2.5",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^8.2.5",
+    "@angular/compiler": "^8.2.5",
+    "@angular/core": "^8.2.5",
+    "@angular/flex-layout": "^7.0.0-beta.24",
+    "@angular/forms": "^8.2.5",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^8.2.5",
+    "@angular/platform-browser-dynamic": "^8.2.5",
+    "@angular/router": "^8.2.5",
+    "@ngrx/effects": "7.2.0",
+    "@ngrx/router-store": "7.2.0",
+    "@ngrx/store": "7.2.0",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.5.4",
     "d3": "^5.7.0",
     "hammerjs": "^2.0.8",
     "openseadragon": "^2.4.0",
-    "rxjs": "6.3.3",
+    "rxjs": "6.5.3",
     "rxjs-compat": "6.0.0",
-    "zone.js": "^0.8.27"
+    "zone.js": "~0.9.1",
+    "@nrwl/angular": "8.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.1",
-    "@angular-devkit/build-ng-packagr": "~0.12.1",
-    "@angular/cli": "7.1.4",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
-    "@ngrx/store-devtools": "7.0.0",
-    "@nrwl/schematics": "7.3.0",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular/cli": "8.0.1",
+    "@angular/compiler-cli": "^8.2.5",
+    "@angular/language-service": "^8.2.5",
+    "@ngrx/store-devtools": "7.2.0",
     "@types/chai": "^4.1.4",
     "@types/chai-as-promised": "^7.1.0",
     "@types/cucumber": "^4.0.4",
@@ -100,7 +99,7 @@
     "angular2-template-loader": "^0.6.2",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "concurrently": "^3.6.0",
     "cucumber": "^5.0.2",
     "jasmine-core": "~2.99.1",
@@ -112,7 +111,7 @@
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "mountebank": "^1.14.1",
-    "ng-packagr": "4.5.0",
+    "ng-packagr": "5.5.0",
     "ngrx-store-freeze": "0.2.4",
     "prettier": "1.15.2",
     "pretty-quick": "^1.8.0",
@@ -122,11 +121,13 @@
     "standard-version": "^4.4.0",
     "ts-node": "~7.0.0",
     "tsickle": "0.34.0",
-    "tslib": "^1.9.0",
+    "tslib": "^1.10.0",
     "tslint": "~5.11.0",
     "tslint-config-prettier": "^1.15.0",
-    "typescript": "3.2.2",
+    "typescript": "~3.4.5",
     "wallaby-webpack": "^3.9.13",
-    "@ngrx/schematics": "7.0.0"
+    "@ngrx/schematics": "7.2.0",
+    "dotenv": "6.2.0",
+    "@nrwl/workspace": "8.0.0"
   }
 }
