@@ -2,7 +2,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { NxModule } from '@nrwl/nx';
+import { NxModule } from '@nrwl/angular';
 import 'hammerjs';
 import 'openseadragon';
 import { AppRoutingModule } from './app-routing.module';
