@@ -4,7 +4,7 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterModule, Routes } from '@angular/router';
-import { NxModule } from '@nrwl/nx';
+import { NxModule } from '@nrwl/angular';
 import 'hammerjs';
 import { MimeModule } from '@nationallibraryofnorway/ngx-mime';
 import { AppComponent } from './app.component';
